# dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/
    description: SwaggerHub API Auto Mocking
...
paths:
...
  /player/playerFullDetails/{playerId}:
    get:
      tags:
        - Players
      summary: get Player personal page by ID
      operationId: PersnoalPlayerPage
      parameters:
        - name: playerId
          in: path
          description: Identification of the requested player by id.
          required: true
          style: simple
          explode: false
          schema:
            type: number
            example: 85764
      responses:
        "200":
          description: OK. recipes on the way.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.      
    # put:
    #   tags:
    #   - Players
    #   summary: update player page
    #   operationId: playerchange
    #   parameters:
    #   - name: playerId
    #     in: path
    #     description: Identification of the requested playr by id.
    #     required: true
    #     style: simple
    #     explode: false
    #     schema:
    #       type: number
    #       example: 234
    #   responses:
    #     "200":
    #       description: success
    #     "400":
    #       description: somting went wrong
  /player/SearchPlayerByName/{playerName}:
    get:
      tags:
        - Players
      summary: Search player by name
      parameters:
        - name: playerName
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: data of specific player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Bad Request.
  /player/SearchPlayerByName/{playerName}/filterbyTeam/{teamName}:
    get:
      tags:
        - Players
      summary: Search player by name
      parameters:
        - name: playerName
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: jesper
        - name: teamName
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: Lyngby
      responses:
        "200":
          description: data of specific player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Bad Request.
  /player/SearchPlayerByName/{playerName}/filterbyPosition/{position}:
    get:
      tags:
        - Players
      summary: Search player by name
      parameters:
        - name: playerName
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: jesper
        - name: position
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
          example: 3
      responses:
        "200":
          description: data of specific player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Bad Request.
# /SearchPlayerByName/:playername/filterbyTeam/:teamName
  /teams/teamFullDetails/{teamId}:
    get:
      tags:
        - Teams
      summary: get team personal page by ID
      operationId: TeamPage
      parameters:
        - name: teamId
          in: path
          description: Identification of the requested team by id.
          required: true
          style: simple
          explode: false
          schema:
            type: number
            example: 85
      responses:
        "200":
          description: OK. recipes on the way.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.
  /teams/teamByName/{team_name}:
    get:
      tags:
        - Teams
      summary: Search team by name
      parameters:
        - name: team_name
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: Horsens

      responses:
        "200":
          description: Data of specific user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Bad Request.
  /homepage/favoriteplayer:
    get:
      tags:
        - Personal Page player
      summary: Get user's favorite players
      responses:
        "200":
          description: favorite team array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_teams"
        "401":
          description: User needs to be logged in
        "400":
          description: somting went wrong
  /homepage/favoriteteam:
    get:
      tags:
        - Personal Page team
      summary: Get user's favorite teams
      responses:
        "200":
          description: favorite team array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_teams"
        "401":
          description: User needs to be logged in
        "400":
          description: somting went wrong
  /homepage/favoriteplayer/{playerId}:
    put:
      tags:
        - Personal Page player
      summary: add a player to the user's favorites
      parameters:
        - name: playerId
          in: path
          required: true
      responses:
        "200":
          description: success
        "400":
          description: somting went wrong
    delete:
      tags:
        - Personal Page player
      summary: Delete a player from the user's favorite players list
      parameters:
        - name: playerId
          in: path
          required: true
      responses:
        "200":
          description: success
        "400":
          description: Bad Request.
  /homepage/favoriteteam/{teamId}:
    put:
      tags:
        - Personal Page team
      summary: add a team to the user's favorites
      parameters:
        - name: teamId
          in: path
          required: true
      responses:
        "200":
          description: success
        "400":
          description: somting went wrong
    delete:
      tags:
        - Personal Page team
      summary: Delete a team from the user's favorite team list
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: success
        "400":
          description: Bad Request.
  /homepage/favoritematches:
    get:
      tags:
        - Personal Page match
      summary: Get user's favorite matches
      responses:
        "200":
          description: favorite match array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_matches"
        "401":
          description: User needs to be logged in
        "400":
          description: somting went wrong
  /homepage/favoritematches/{gameId}:
    delete:
      tags:
        - Personal Page match
      summary: Delete a game from the user's favorite game list
      parameters:
        - name: gameId
          in: path
          required: true

      responses:
        "200":
          description: success
        "400":
          description: Bad Request.
    put:
      tags:
        - Personal Page match
      summary: add a match to the user's favorites
      parameters:
      - name: gameId
        in: path
        required: true
      responses:
        "200":
          description: success
        "400":
          description: somting went wrong  
...
components:
  schemas:
...
    Referee:
      type: object
      properties:
        id:
          type: integer
          example: 33
        Name:
          type: string
          example: Maccabi Tel Aviv
        image-url:
          type: string
          format: URL
        qualification:
          type: string
          example: Line max
...
    Player:
      type: object
      properties:
        id:
          type: integer
          example: 82808
        fullName:
          type: string
          example: Jesper Hansen
        teamName:
          example: Midtjylland
        image-url:
          type: string
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/24/82808.png
        position:
          type: integer
          example: 1
...
    Team:
      type: object
      properties:
        id:
          type: integer
          example: 85
        teamName:
          type: string
          example: København
        coach-id:
          type: integer
          example: 456421
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        Games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
    Game:
      type: object
      properties:
        gameDate:
          type: string
          example: 19941030
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        referees:
          $ref: "#/components/schemas/Referee"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        homeScore:
          type: integer
          example: 2
        awayScore:
          example: 3
    Event:
      type: object
      properties:
        gameId:
          type: integer
          example: 12
        eventType:
          type: string
          example: "red Card"
        gameDate:
          type: string
          example: 19941030
        gameTime:
          type: string
          example: "17:30:25"
        inGameMinute:
          type: integer
          example: 13
        eventDescription:
          type: string
          example: goal by Ronaldo
...